Compose area: Workaround for placeholder not showing up sometimes
Steps to reproduce:

1. Type some text in the compose area
2. Highlight it all and delete it (ctrl+a, ctrl+x)

The previous behavior happened because the contenteditable still
contained non-visible HTML content (an empty <div>), which means it's
technically not empty and therefore won't display the placeholder.

This bug was not present in older versions and started appearing with an
Electron update. Most probably Chromium changed the way contenteditable
elements are handled.