Fix crash due to "broken pipe" on Linux
Automated restarts through the launcher binary on Linux were broken
since beta 45 when started through the `.desktop` file (and not on the
terminal) and caused a crash due to SIGPIPE (resulting in a SIGSEGV of
the flatpak).

The reason is that mistakenly data was printed to stdout even when no
stdout is allocated (which is the case when starting a Linux application
through an app launcher). And the Rust printing macros unwrap/panic when
stdout is not available. In theory we have macros to prevent that, but
the macro mistakenly checked for stderr, and not stdout. And
unfortunately stderr *is* available in that case, just not stdout 🤦